# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pd-prep-for-pgdp",
    "description": "Convert a folder/zip of scanned book images into a PGDP-ready submission package.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "List Recent Jobs",
        "operationId": "list_recent_jobs",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Recent Jobs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/gpu/jobs": {
      "get": {
        "tags": [
          "gpu"
        ],
        "summary": "List Jobs",
        "description": "List the most recent jobs for the current user (newest first).",
        "operationId": "list_gpu_jobs",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response List Gpu Jobs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "gpu"
        ],
        "summary": "Submit Batch Job",
        "operationId": "submit_batch_job",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchJobRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gpu/jobs/{job_id}": {
      "get": {
        "tags": [
          "gpu"
        ],
        "summary": "Get Job",
        "operationId": "get_gpu_job",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "gpu"
        ],
        "summary": "Cancel Job",
        "operationId": "cancel_gpu_job",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gpu/jobs/{job_id}/retry": {
      "post": {
        "tags": [
          "gpu"
        ],
        "summary": "Retry Job",
        "description": "Create a fresh copy of a failed/cancelled job in `queued` status.\n\nSame project_id and type \u2014 fresh id and timestamps. The original job\nstays in the database so the user can compare. The new job's payload\nstarts as a copy of the original's; if `body.payload_override` is\nprovided (P3 #16), its keys are shallow-merged over the original\npayload (override keys replace, others are preserved). Pass `None` /\nomit the body to retry verbatim.",
        "operationId": "retry_gpu_job",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RetryJobRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchJobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/gpu/jobs/{job_id}/events": {
      "get": {
        "tags": [
          "gpu"
        ],
        "summary": "Job Events",
        "description": "SSE \u2014 push status transitions until the job is terminal.\n\nSubscribes to the in-process `JobEventBroker` and forwards events as\n`text/event-stream` frames. First frame is the current snapshot from\nthe database so a late subscriber sees state immediately; subsequent\nframes come from the broker (zero-poll).",
        "operationId": "stream_gpu_job_events",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BatchJobRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "job_type": {
            "type": "string",
            "enum": [
              "batch_process_pages",
              "batch_ocr",
              "batch_text_postprocess",
              "batch_extract_illustrations",
              "build_package"
            ],
            "title": "Job Type"
          },
          "page_idxs": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Idxs"
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "job_type"
        ],
        "title": "BatchJobRequest"
      },
      "BatchJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "estimated_pages": {
            "type": "integer",
            "title": "Estimated Pages",
            "default": 0
          },
          "dispatch_mode": {
            "type": "string",
            "enum": [
              "immediate",
              "scheduled"
            ],
            "title": "Dispatch Mode",
            "default": "immediate"
          },
          "next_dispatch_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Dispatch At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "BatchJobResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": "default"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus",
            "default": "queued"
          },
          "progress": {
            "$ref": "#/components/schemas/JobProgress"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "next_dispatch_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Dispatch At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "gpu_backend": {
            "type": "string",
            "enum": [
              "local",
              "cpu",
              "modal",
              "shared_container"
            ],
            "title": "Gpu Backend",
            "default": "local"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "owner_id",
          "type",
          "status",
          "progress",
          "created_at",
          "started_at",
          "completed_at",
          "next_dispatch_at",
          "error_message",
          "gpu_backend",
          "payload"
        ],
        "title": "Job"
      },
      "JobProgress": {
        "properties": {
          "current": {
            "type": "integer",
            "title": "Current",
            "default": 0
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "current_page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Page"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "current",
          "total",
          "current_page",
          "message"
        ],
        "title": "JobProgress"
      },
...
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "scheduled",
          "running",
          "complete",
          "error",
          "cancelled"
        ],
        "title": "JobStatus"
      },
      "JobType": {
        "type": "string",
        "enum": [
          "unzip",
          "thumbnails",
          "batch_process_pages",
          "batch_ocr",
          "batch_text_postprocess",
          "batch_extract_illustrations",
          "build_package",
          "run_page_stage"
        ],
        "title": "JobType"
      },
...
      "PipelineState": {
        "properties": {
          "steps": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StepState"
            },
            "type": "object",
            "title": "Steps"
          }
        },
        "type": "object",
        "required": [
          "steps"
        ],
        "title": "PipelineState"
      },
...
      "Project": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_id": {
            "type": "string",
            "title": "Owner Id",
            "default": "default"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatus"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "proof_page_count": {
            "type": "integer",
            "title": "Proof Page Count"
          },
          "config": {
            "$ref": "#/components/schemas/ProjectConfig"
          },
          "pipeline_state": {
            "$ref": "#/components/schemas/PipelineState"
          },
          "storage_prefix": {
            "type": "string",
            "title": "Storage Prefix"
          },
          "archived": {
            "type": "boolean",
            "title": "Archived",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "name",
          "created_at",
          "updated_at",
          "status",
          "page_count",
          "proof_page_count",
          "config",
          "pipeline_state",
          "storage_prefix",
          "archived"
        ],
        "title": "Project"
      },
      "ProjectConfig": {
        "properties": {
          "book_name": {
            "type": "string",
            "title": "Book Name"
          },
          "source_uri": {
            "type": "string",
            "title": "Source Uri"
          },
          "proof_start_idx0": {
            "type": "integer",
            "title": "Proof Start Idx0",
            "default": 0
          },
          "proof_end_idx0": {
            "type": "integer",
            "title": "Proof End Idx0",
            "default": 0
          },
          "cover_idx0": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Idx0"
          },
          "title_idx0": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title Idx0"
          },
          "frontmatter_start_idx0": {
            "type": "integer",
            "title": "Frontmatter Start Idx0",
            "default": 0
          },
          "frontmatter_end_idx0": {
            "type": "integer",
            "title": "Frontmatter End Idx0",
            "default": 0
          },
          "bodymatter_start_idx0": {
            "type": "integer",
            "title": "Bodymatter Start Idx0",
            "default": 0
          },
          "bodymatter_end_idx0": {
            "type": "integer",
            "title": "Bodymatter End Idx0",
            "default": 0
          },
          "frontmatter_page_nbr_start": {
            "type": "integer",
            "title": "Frontmatter Page Nbr Start",
            "default": 1
          },
          "bodymatter_page_nbr_start": {
            "type": "integer",
            "title": "Bodymatter Page Nbr Start",
            "default": 1
          },
          "initial_crop_all": {
            "prefixItems": [
              {
                "type": "integer"
              },
              {
                "type": "integer"
              },
              {
                "type": "integer"
              },
              {
                "type": "integer"
              }
            ],
            "type": "array",
            "maxItems": 4,
            "minItems": 4,
            "title": "Initial Crop All",
            "default": [
              0,
              0,
              0,
              0
            ]
          },
          "ocr_crop_top": {
            "type": "integer",
            "title": "Ocr Crop Top",
            "default": 0
          },
          "ocr_crop_bottom": {
            "type": "integer",
            "title": "Ocr Crop Bottom",
            "default": 0
          },
          "ocr_crop_left": {
            "type": "integer",
            "title": "Ocr Crop Left",
            "default": 0
          },
          "ocr_crop_right": {
            "type": "integer",
            "title": "Ocr Crop Right",
            "default": 0
          },
          "custom_regex_passes": {
            "items": {
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array",
              "maxItems": 2,
              "minItems": 2
            },
            "type": "array",
            "title": "Custom Regex Passes"
          },
          "custom_scannos": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Custom Scannos"
          },
          "layout_category_overrides": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object",
            "title": "Layout Category Overrides"
          },
          "default_overrides": {
            "additionalProperties": true,
            "type": "object",
            "title": "Default Overrides"
          }
        },
        "type": "object",
        "required": [
          "book_name",
          "source_uri",
          "proof_start_idx0",
          "proof_end_idx0",
          "cover_idx0",
          "title_idx0",
          "frontmatter_start_idx0",
          "frontmatter_end_idx0",
          "bodymatter_start_idx0",
          "bodymatter_end_idx0",
          "frontmatter_page_nbr_start",
          "bodymatter_page_nbr_start",
          "initial_crop_all",
          "ocr_crop_top",
          "ocr_crop_bottom",
          "ocr_crop_left",
          "ocr_crop_right",
          "custom_regex_passes",
          "custom_scannos",
          "layout_category_overrides",
          "default_overrides"
        ],
        "title": "ProjectConfig"
      },
      "ProjectStatus": {
        "type": "string",
        "enum": [
          "ingesting",
          "configuring",
          "processing",
          "reviewing",
          "packaging",
          "complete"
        ],
        "title": "ProjectStatus"
      },
      "RetryJobRequest": {
        "properties": {
          "payload_override": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Override"
          }
        },
        "type": "object",
        "title": "RetryJobRequest",
        "description": "Optional body for `POST /api/gpu/jobs/{id}/retry`.\n\n`payload_override`, when non-null, is shallow-merged over the original\njob's payload \u2014 keys present in the override replace the corresponding\nkeys; keys not present are preserved from the original. The original\njob is never mutated, so the audit trail stays intact."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}